fc_util: some refactoring
1. Move time related functions to a separate
   module.
2. Create a new module called rand and move rand
   related functionality like `xor_rng_u32` there.
