+++ api_server/src/lib.rs
@@ -196,7 +196,8 @@ impl ApiServer {
         let listener = UnixListener::bind(path, &handle).map_err(Error::Io)?;
 
         if let Some(start_time) = start_time_us {
-            let delta_us = (fc_util::get_time(fc_util::ClockType::Monotonic) / 1000) - start_time;
+            let delta_us =
+                (fc_util::time::get_time(fc_util::time::ClockType::Monotonic) / 1000) - start_time;
             METRICS
                 .api_server
                 .process_startup_time_us
@@ -204,8 +205,8 @@ impl ApiServer {
         }
 
         if let Some(cpu_start_time) = start_time_cpu_us {
-            let delta_us =
-                fc_util::get_time(fc_util::ClockType::ProcessCpu) / 1000 - cpu_start_time;
+            let delta_us = fc_util::time::get_time(fc_util::time::ClockType::ProcessCpu) / 1000
+                - cpu_start_time;
             METRICS
                 .api_server
                 .process_startup_time_cpu_us
